requirements.txt
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.8
attrs==25.1.0
autoflake==2.3.1
backports.tarfile==1.2.0
black==24.10.0
blis==1.3.0
boto3==1.37.32
botocore==1.37.32
build==1.2.2.post1
CacheControl==0.14.2
catalogue==2.0.10
Cerberus==1.3.7
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
coverage==7.6.12
crashtest==0.4.1
crispy-bootstrap5==0.7
cryptography==44.0.0
cssbeautifier==1.15.4
cymem==2.0.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.6
django-appconf==1.1.0
django-bunny==1.1.6
django-cockroachdb==5.1
django-configurations==2.5.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-discover-runner==1.0
django-filter==23.5
django-mailer==2.3.2
django-nose==1.4.7
django-prometheus==2.3.1
django-recaptcha==3.0.0
django-redis==5.4.0
django-select2==8.4.1
django-storages==1.14.4
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-dataclasses==1.3.1
djangorestframework_simplejwt==5.4.0
djlint==1.36.4
djoser==2.3.1
drf-extensions==0.7.1
drf-spectacular==0.26.5
drf-spectacular-sidecar==2024.12.1
dulwich==0.22.7
EditorConfig==0.17.1
fastcore==1.7.29
fastjsonschema==2.21.1
filelock==3.17.0
fsspec==2025.7.0
ghapi==1.0.6
googleapis-common-protos==1.66.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==22.0.0
h11==0.14.0
hf-xet==1.1.5
highlight-io==0.8.10
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.34.3
humanize==4.12.1
identify==2.6.7
idna==3.10
importlib_metadata==8.0.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.12.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kolo==2.37.0
langcodes==3.5.0
language_data==1.3.0
lockfile==0.12.2
logtail-python==0.2.10
loguru==0.7.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mocket==3.13.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.13
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
nodeenv==1.9.1
nose==1.3.7
numpy==2.2.6
oauthlib==3.2.2
opentelemetry-api==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-anthropic==0.26.5
opentelemetry-instrumentation-bedrock==0.26.5
opentelemetry-instrumentation-boto==0.47b0
opentelemetry-instrumentation-boto3sqs==0.47b0
opentelemetry-instrumentation-celery==0.47b0
opentelemetry-instrumentation-chromadb==0.26.5
opentelemetry-instrumentation-cohere==0.26.5
opentelemetry-instrumentation-haystack==0.26.5
opentelemetry-instrumentation-langchain==0.26.5
opentelemetry-instrumentation-llamaindex==0.26.5
opentelemetry-instrumentation-logging==0.47b0
opentelemetry-instrumentation-openai==0.26.5
opentelemetry-instrumentation-pinecone==0.26.5
opentelemetry-instrumentation-qdrant==0.26.5
opentelemetry-instrumentation-redis==0.47b0
opentelemetry-instrumentation-replicate==0.26.5
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0
opentelemetry-instrumentation-transformers==0.26.5
opentelemetry-instrumentation-vertexai==0.26.5
opentelemetry-instrumentation-watsonx==0.26.5
opentelemetry-instrumentation-weaviate==0.26.5
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.1
opentelemetry-util-http==0.47b0
packaging==24.2
pathspec==0.12.1
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
poetry-plugin-export==1.9.0
pre-commit==3.8.0
preshed==3.0.10
prometheus_client==0.21.1
promise==2.3
protobuf==4.25.6
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
puremagic==1.28
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.22.3
ruff==0.12.7
s3transfer==0.11.4
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sourcery==1.33.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlglot==26.6.0
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
text-unidecode==1.3
thinc==8.3.6
tiktoken==0.8.0
tokenizers==0.21.4
tomlkit==0.13.2
toolz==1.0.0
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
trove-classifiers==2025.1.15.22
typer==0.16.0
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
ulid-py==1.1.0
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.2
wasabi==1.1.3
weasel==0.4.1
whitenoise==6.9.0
wrapt==1.17.2
xattr==1.1.4
zipp==3.21.0
